refactor(profiling): extract Clocks
