+++ profiling/src/profiling/mod.rs
@@ -612,25 +612,6 @@ impl Profiler {
         }
     }
 
-    fn cpu_sub(now: cpu_time::ThreadTime, prev: cpu_time::ThreadTime) -> i64 {
-        let now = now.as_duration();
-        let prev = prev.as_duration();
-
-        match now.checked_sub(prev) {
-            // If a 128 bit value doesn't fit in 64 bits, use the max.
-            Some(duration) => duration.as_nanos().try_into().unwrap_or(i64::MAX),
-
-            // If this happened, then either the programmer screwed up and
-            // passed args in backwards, or cpu time has gone backward... ish.
-            // Supposedly it can happen if the thread migrates CPUs:
-            // https://www.percona.com/blog/what-time-18446744073709550000-means/
-            // Regardless of why, a customer hit this:
-            // https://github.com/DataDog/dd-trace-php/issues/1880
-            // In these cases, zero is much closer to reality than i64::MAX.
-            None => 0,
-        }
-    }
-
     /// Collect a stack sample with elapsed wall time. Collects CPU time if
     /// it's enabled and available.
     pub unsafe fn collect_time(
@@ -645,28 +626,7 @@ impl Profiler {
         match result {
             Ok(frames) => {
                 let depth = frames.len();
-                let (wall_time, cpu_time) = CLOCKS.with(|cell| {
-                    let mut tls_clocks = cell.borrow_mut();
-
-                    let wall_now = Instant::now();
-                    let wall_time = wall_now.duration_since(tls_clocks.wall_time);
-                    tls_clocks.wall_time = wall_now;
-                    let wall_time: i64 = wall_time.as_nanos().try_into().unwrap_or(i64::MAX);
-
-                    // If CPU time is disabled, or if it's enabled but not available on the
-                    // platform, then `tsl_clocks.cpu_time` will be None.
-                    let cpu_time = if let Some(last_cpu_time) = tls_clocks.cpu_time {
-                        let now = cpu_time::ThreadTime::try_now().expect(
-                            "CPU time to work since it's worked before during this process",
-                        );
-                        let cpu_time = Self::cpu_sub(now, last_cpu_time);
-                        tls_clocks.cpu_time = Some(now);
-                        cpu_time
-                    } else {
-                        0
-                    };
-                    (wall_time, cpu_time)
-                });
+                let (wall_time, cpu_time) = CLOCKS.with(|cell| cell.borrow_mut().rotate_clocks());
 
                 let mut labels = Profiler::message_labels();
                 #[cfg(feature = "timeline")]
